requirements.txt
Flask==2.2.5
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
python-dotenv==1.0.0
pytz==2023.3.post1
Werkzeug==2.2.2
